requirements.txt
fabric
python-dotenv
requests
